finance-assistant/requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
openai==1.3.7
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0
plotly==5.17.0
pandasql==0.7.3